shop details view rearangement
