+++ app/src/main/res/layout/fragment_shop_details.xml
@@ -23,39 +23,44 @@
         android:id="@+id/shop_name_view"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:text="TextView"
+        android:text="Dworzec Główny, Poznań"
         android:textAlignment="center"
         android:textAppearance="@style/TextAppearance.AppCompat.Large"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/shop_details_map_view" />
+        app:layout_constraintTop_toBottomOf="@+id/start_navigation_button" />
 
     <TextView
-        android:id="@+id/shop_features_view"
+        android:id="@+id/shop_address_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="TextView"
+        android:text="ul. Dworcowa 12, 60-456 Poznań"
+        android:textAppearance="@style/TextAppearance.AppCompat.Body2"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/shop_name_view" />
-
-    <TextView
-        android:id="@+id/shop_open_hours_view"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Open Hours"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/shop_features_view" />
 
     <TextView
         android:id="@+id/shop_distance_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="912 m away"
+        android:textAppearance="@style/TextAppearance.AppCompat.Body2"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/shop_address_view" />
+
+    <TextView
+        android:id="@+id/shop_open_hours_view"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Pn-Pt - 6:00 - 22:00\nSo - 08:00:22:00\nNd - Nieczynne"
+        android:textAppearance="@style/TextAppearance.AppCompat.Body1"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/shop_open_hours_view" />
+        app:layout_constraintTop_toBottomOf="@+id/shop_distance_view" />
 
     <Button
         android:id="@+id/start_navigation_button"
@@ -66,6 +71,16 @@
         android:text="@string/navigate_to_shop"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/shop_distance_view" />
+        app:layout_constraintTop_toBottomOf="@+id/shop_details_map_view" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/shop_features_view"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/shop_open_hours_view" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
