+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -16,10 +16,6 @@
         <action
             android:id="@+id/action_navigation_list_to_navigation_shop_details"
             app:destination="@id/navigation_shop_details"
-            app:enterAnim="@anim/fragment_open_enter"
-            app:exitAnim="@anim/fragment_open_exit"
-            app:popEnterAnim="@anim/fragment_fade_enter"
-            app:popExitAnim="@anim/fragment_fade_exit"
             app:popUpTo="@id/navigation_list" />
     </fragment>
