+++ buildSrc/src/main/java/Dependencies.kt
@@ -1,5 +1,5 @@
 object Versions {
-    const val kotlin = "1.4.10"
+    const val kotlin = "1.4.20"
     const val kotlinSerialization = "1.0.0."
     const val ktor: String = "1.4.1"
     const val koin = "2.2.0-beta-2"
