+++ buildSrc/src/main/java/ProjectDeps.kt
@@ -1,10 +1,10 @@
 object ProjectDeps {
-  const val gradlePlugin = "com.android.tools.build:gradle:4.1.0"
-  const val googleServices = "com.google.gms:google-services:4.3.4"
-  const val kotlinPlugin = "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.kotlin}"
+    const val gradlePlugin = "com.android.tools.build:gradle:4.1.0"
+    const val googleServices = "com.google.gms:google-services:4.3.4"
+    const val kotlinPlugin = "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.kotlin}"
     const val kotlinSerialization = "org.jetbrains.kotlin:kotlin-serialization:${Versions.kotlin}"
     const val kotlinExtensions = "org.jetbrains.kotlin:kotlin-android-extensions:${Versions.kotlin}"
-  const val firebasePlugin = "com.google.firebase:firebase-plugins:1.0.5"
-  const val dokka = "org.jetbrains.dokka:dokka-android-gradle-plugin:0.9.17"
-  const val crashlytics = "com.google.firebase:firebase-crashlytics-gradle:2.3.0"
+    const val firebasePlugin = "com.google.firebase:firebase-plugins:1.0.5"
+    const val dokka = "org.jetbrains.dokka:dokka-android-gradle-plugin:0.9.17"
+    const val crashlytics = "com.google.firebase:firebase-crashlytics-gradle:2.3.0"
 }
